问题调整

+++ homework/src/main/java/com/jlm/homework/repository/BookKnowledgePointRepository.java
@@ -3,5 +3,5 @@ package com.jlm.homework.repository;
 import com.jlm.homework.entity.BookKnowledgePoint;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookKnowledgePointRepository extends JpaRepository<BookKnowledgePoint, Integer> {
+public interface BookKnowledgePointRepository extends JpaRepository<BookKnowledgePoint, Long> {
 }
