+++ homework/src/main/java/com/jlm/homework/repository/ExerciseBookChapterRepository.java
@@ -3,5 +3,5 @@ package com.jlm.homework.repository;
 import com.jlm.homework.entity.ExerciseBookChapter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExerciseBookChapterRepository extends JpaRepository<ExerciseBookChapter,Integer> {
+public interface ExerciseBookChapterRepository extends JpaRepository<ExerciseBookChapter,Long> {
 }
